+++ src/internals/swr-queue.ts
@@ -20,7 +20,7 @@
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
 import Queue from 'bee-queue'
-import { option as O } from 'fp-ts'
+import { either as E, option as O } from 'fp-ts'
 import * as R from 'ramda'
 
 import { Cache, Priority } from './cache'
@@ -33,6 +33,7 @@ import { modelFactories } from '~/model'
 export interface SwrQueue {
   queue(updateJob: UpdateJob): Promise<never>
   ready(): Promise<void>
+  healthy(): Promise<void>
   quit(): Promise<void>
   _queue: never
 }
@@ -48,6 +49,9 @@ export const emptySwrQueue: SwrQueue = {
   ready() {
     return Promise.resolve()
   },
+  healthy() {
+    return Promise.resolve()
+  },
   quit() {
     return Promise.resolve()
   },
@@ -56,7 +60,24 @@ export const emptySwrQueue: SwrQueue = {
 
 export const queueName = 'swr'
 
-export function createSwrQueue(): SwrQueue {
+export function createSwrQueue({
+  cache,
+  timer,
+}: {
+  cache: Cache
+  timer: Timer
+}): SwrQueue {
+  const args = {
+    environment: {
+      cache,
+      swrQueue: emptySwrQueue,
+    },
+    fetchHelpers: createFetchHelpersFromNodeFetch(),
+  }
+  const models = R.values(modelFactories).map((createModel) =>
+    createModel(args)
+  )
+
   const queue = new Queue<UpdateJob>(queueName, {
     redis: {
       url: redisUrl,
@@ -69,7 +90,22 @@ export function createSwrQueue(): SwrQueue {
   return {
     _queue: (queue as unknown) as never,
     async queue(updateJob) {
-      log.debug('Queuing job', updateJob.key)
+      const { key } = updateJob
+
+      const result = await shouldProcessJob({
+        key,
+        cache,
+        models,
+        timer,
+      })
+
+      if (E.isLeft(result)) {
+        log.debug('Skipped job', key, 'because', result.left)
+        return undefined as never
+      }
+
+      log.debug('Queuing job', key)
+
       // By setting the job's ID, we make sure that there will be only one update job for the same key
       // See also https://github.com/bee-queue/bee-queue#jobsetidid
       const job = await queue.createJob(updateJob).setId(updateJob.key).save()
@@ -93,6 +129,9 @@ export function createSwrQueue(): SwrQueue {
     async ready() {
       await queue.ready()
     },
+    async healthy() {
+      await queue.checkHealth()
+    },
     async quit() {
       await queue.close()
     },
@@ -109,6 +148,7 @@ export function createSwrQueueWorker({
   concurrency: number
 }): {
   ready(): Promise<void>
+  healthy(): Promise<void>
   quit(): Promise<void>
   _queue: never
 } {
@@ -131,44 +171,28 @@ export function createSwrQueueWorker({
     removeOnSuccess: true,
   })
 
-  function getSpec(key: string): QuerySpec<unknown, unknown> | null {
-    for (const model of models) {
-      for (const prop of Object.values(model)) {
-        if (isQuery(prop) && O.isSome(prop._querySpec.getPayload(key))) {
-          return prop._querySpec
-        }
-      }
-    }
-    return null
-  }
-
   queue.process(
     concurrency,
     async (job): Promise<string> => {
       const { key } = job.data
-      const cacheEntry = await cache.get<unknown>({ key })
-      if (O.isNone(cacheEntry)) {
-        return 'Skipped update because cache empty.'
+
+      const result = await shouldProcessJob({
+        key,
+        cache,
+        models,
+        timer,
+      })
+      if (E.isLeft(result)) {
+        return `Skipped update because ${result.left}`
       }
-      const spec = getSpec(key)
-      if (spec === null) {
-        return 'Skipped update because invalid key.'
-      }
-      const maxAge =
-        spec.maxAge === undefined ? undefined : timeToMilliseconds(spec.maxAge)
-      const age = timer.now() - cacheEntry.value.lastModified
-      if (maxAge === undefined || age <= maxAge) {
-        return 'Skipped update because cache non-stale.'
-      }
-      const payload = spec.getPayload(key)
-      if (O.isNone(payload)) {
-        return 'Skipped updated because invalid key.'
-      }
+
+      const { spec, payload } = result.right
+
       await cache.set({
         key,
         priority: Priority.Low,
         getValue: async (current) => {
-          return await spec.getCurrentValue(payload.value, current ?? null)
+          return await spec.getCurrentValue(payload, current ?? null)
         },
       })
       return 'Updated because stale'
@@ -180,10 +204,65 @@ export function createSwrQueueWorker({
     async ready() {
       await queue.ready()
     },
+    async healthy() {
+      await queue.checkHealth()
+    },
     async quit() {
       await queue.close()
     },
   }
+}
+
+async function shouldProcessJob({
+  key,
+  cache,
+  models,
+  timer,
+}: {
+  key: string
+  cache: Cache
+  models: Record<string, unknown>[]
+  timer: Timer
+}): Promise<
+  E.Either<string, { spec: QuerySpec<unknown, unknown>; payload: unknown }>
+> {
+  function getSpec(key: string): QuerySpec<unknown, unknown> | null {
+    for (const model of models) {
+      for (const prop of Object.values(model)) {
+        if (isQuery(prop) && O.isSome(prop._querySpec.getPayload(key))) {
+          return prop._querySpec
+        }
+      }
+    }
+    return null
+  }
+
+  const cacheEntry = await cache.get<unknown>({ key })
+  if (O.isNone(cacheEntry)) {
+    return E.left('cache empty.')
+  }
+  const spec = getSpec(key)
+  if (spec === null) {
+    return E.left('invalid key.')
+  }
+  if (!spec.enableSwr) {
+    return E.left('SWR disabled.')
+  }
+  const maxAge =
+    spec.maxAge === undefined ? undefined : timeToMilliseconds(spec.maxAge)
+  const age = timer.now() - cacheEntry.value.lastModified
+  if (maxAge === undefined || age <= maxAge) {
+    return E.left('cache non-stale.')
+  }
+  const payload = spec.getPayload(key)
+  if (O.isNone(payload)) {
+    return E.left('invalid key.')
+  }
+
+  return E.right({
+    spec,
+    payload: payload.value,
+  })
 }
 
 export interface Time {
@@ -209,8 +288,8 @@ export function timeToMilliseconds({
 }: Time) {
   const SECOND = 1000
   const MINUTE = 60 * SECOND
-  const HOUR = 60 * minute
-  const DAY = 24 * hour
+  const HOUR = 60 * MINUTE
+  const DAY = 24 * HOUR
 
   return (
     (day + days) * DAY +