+++ CHANGELOG.md
@@ -2,7 +2,67 @@
 
 All notable changes to this project will be documented in this file.
 
-## [v0.16.0](https://github.com/serlo/api.serlo.org/compare/v0.15.0..v0.16.0) - January 6, 2021
+## [v0.17.7](https://github.com/serlo/api.serlo.org/compare/v0.17.6..v0.17.7) - January 21, 2021
+
+### Fixed
+
+- **swr-queue**. Fix `Time`.
+
+## [v0.17.6](https://github.com/serlo/api.serlo.org/compare/v0.17.5..v0.17.6) - January 20, 2021
+
+### Internal
+
+- **cache**. Adapt `maxAge`.
+
+## [v0.17.5](https://github.com/serlo/api.serlo.org/compare/v0.17.4..v0.17.5) - January 19, 2021
+
+### Fixed
+
+- **cache**. Handle cache values without timestamp correctly.
+
+## [v0.17.4](https://github.com/serlo/api.serlo.org/compare/v0.17.3..v0.17.4) - January 16, 2021
+
+### Internal
+
+- serlo-org-database-layer@0.1.6
+
+## [v0.17.3](https://github.com/serlo/api.serlo.org/compare/v0.17.2..v0.17.3) - January 15, 2021
+
+### Internal
+
+- serlo-org-database-layer@0.1.5
+
+- Split contract into serlo.org and serlo.org-database-layer.
+
+## [v0.17.2](https://github.com/serlo/api.serlo.org/compare/v0.17.1..v0.17.2) - January 14, 2021
+
+### Internal
+
+- serlo-org-database-layer@0.1.4
+
+- Add health checks
+
+## [v0.17.1](https://github.com/serlo/api.serlo.org/compare/v0.17.0..v0.17.1) - January 14, 2021
+
+### Internal
+
+- serlo-org-database-layer@0.1.2
+
+## [v0.17.0](https://github.com/serlo/api.serlo.org/compare/v0.16.1..v0.17.0) - January 11, 2021
+
+### Breaking Changes
+
+- **uuid**. Move `threads` from `AbstractUuid` into concrete types.
+
+### Added
+
+- Add `ThreadAware` interface for types with `threads`.
+
+### Internal
+
+- Add preparations for database layer.
+
+## [v0.16.1](https://github.com/serlo/api.serlo.org/compare/v0.16.0..v0.16.1) - January 6, 2021
 
 ### Breaking Changes
 
@@ -17,6 +77,8 @@ All notable changes to this project will be documented in this file.
 - Split up `api` docker image into `api-server` and `api-swr-queue-worker`.
 
 - Overhaul naming of environment variables.
+
+## [v0.16.0](https://github.com/serlo/api.serlo.org/compare/v0.15.0..v0.16.0) - January 6, 2021 \[YANKED]
 
 ## [v0.15.0](https://github.com/serlo/api.serlo.org/compare/v0.14.2..v0.15.0) - January 5, 2021